step3/requirements.txt
azure-cognitiveservices-vision-customvision==3.1.0
azure-cognitiveservices-vision-face==0.6.0
azure-common==1.1.28
azure-core==1.29.5
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.8.0
urllib3==2.0.7
video-indexer==0.1.8
